packages/airflow/pyproject.toml
[project]
name = "airflow"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4.0"

dependencies = [
    "apache-airflow==3.0.4",
    "psycopg[binary,pool]",
    "psycopg2-binary",
    "apache-airflow-providers-postgres>=6.2.3",
]

[tool.uv]
dev-dependencies = [
    "pipdeptree==2.26.1",
]
